// src/app/empleados/interfaces/empleado.ts
export interface IEmpleado{
    nombres: string;
    apellidos: string;
    direccion:string;
    sueldo: number;
}



// [{{ nombres:'Juan',
// apellidos:'Perez',
// direccion:'Los angles 65', 
// sueldo: 3000,
// },

// { nombres:'Carlos',
// apellidos:'Rojas',
// direccion:'Purto siciali', 
// sueldo: 4000,
// },

// { nombres:'Ramon',
// apellidos:'Valdes',
// direccion:'Purto Mexico', 
// sueldo: 5000,
// },

// }]